Breaking change: new config file format.
This is a lot more succint and easier to read than JSON,
allows comments, and is easily extendible in the future.

Plugins are now lazy loaded as checks/alerts are found
that require them. Closes #18.

This also simplifies the way plugins are defined (skipping the
intermediate CheckType and AlertType structs, and not mandating
a separate struct for configuration).