Parallelise check running.
Scheduled checks are now written to a channel, where job workers
can pick them up. This allows multiple checks to run simultaneously.

Checks are now scheduled based on when they finish - if a check
has an interval of 1 second and takes 10 seconds to run, it would
effectively run every 11 seconds.

Checks that panic (due to errors in plugins) are how handled and
cause a failure.