Small refactoring around check status + configs
Move the config handling out of the Plum struct as it's independent,
and make it handle multiple targets and tracking of keys that are
unused.

Also add listeners for when a check is completed, and use those
for the existing status change functionality. This is making way
for exposing events over an API (#30).