Validate check's alert configs.
If a check has alerts configured then each entry must match at least
one alert for it to be considered valid. This validation is performed
at startup and Goplum will immediately exit if any checks have bad
alerts.

Previously if you typo'd an alert name, Goplum printed an error in
its log when it tried to raise the alert which is too late to be
useful.

The special alert name "-" is allowed in check rules for the case
where you wish to suppress all alerting for a check.

Closes #36